Add decimal support for water intake and goal tracking with safe numeric migration
- migrate intake and goal columns from integer to numeric(6,1) in Supabase
- preserve existing values during migration and keep schema definition in sync
- remove integer rounding in server actions and normalize writes to one decimal place
- allow 0.1-step input for intake and goals, including sub-1 oz entries
- coerce numeric query/realtime payloads to numbers for reliable calculations
- update amount formatting to display clean decimal values across dashboard and history

+++ supabase/schema.sql
@@ -35,7 +35,7 @@ CREATE TABLE public.users (
   id           UUID        PRIMARY KEY REFERENCES auth.users(id) ON DELETE CASCADE,
   email        TEXT        NOT NULL,
   display_name TEXT,
-  daily_goal   INTEGER     NOT NULL DEFAULT 32,
+  daily_goal   NUMERIC(6,1) NOT NULL DEFAULT 32.0,
   avatar_url   TEXT,
   created_at   TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
@@ -82,7 +82,7 @@ CREATE TABLE public.intake_logs (
   id         UUID        PRIMARY KEY DEFAULT gen_random_uuid(),
   user_id    UUID        NOT NULL REFERENCES auth.users(id) ON DELETE CASCADE,
   date       DATE        NOT NULL DEFAULT current_date,
-  ounces     INTEGER     NOT NULL CHECK (ounces > 0),
+  ounces     NUMERIC(6,1) NOT NULL CHECK (ounces > 0),
   created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
   updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
@@ -204,7 +204,7 @@ CREATE TABLE public.daily_goal_overrides (
   id         UUID        PRIMARY KEY DEFAULT gen_random_uuid(),
   user_id    UUID        NOT NULL REFERENCES auth.users(id) ON DELETE CASCADE,
   date       DATE        NOT NULL,
-  daily_goal INTEGER     NOT NULL CHECK (daily_goal > 0),
+  daily_goal NUMERIC(6,1) NOT NULL CHECK (daily_goal > 0),
   created_at TIMESTAMPTZ NOT NULL DEFAULT NOW(),
   UNIQUE(user_id, date)
 );
